+++ control-plane/api/v1alpha1/peeringacceptor_types.go
@@ -1,10 +1,16 @@
 package v1alpha1
 
 import (
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apimachinery/pkg/util/validation/field"
 )
 
 // NOTE: json tags are required.  Any new fields you add must have json tags for the fields to be serialized.
+
+const PeeringAcceptorKubeKind = "peeringacceptors"
+const SecretBackendTypeKubernetes = "kubernetes"
 
 func init() {
 	SchemeBuilder.Register(&PeeringAcceptor{}, &PeeringAcceptorList{})
@@ -85,4 +91,36 @@ func (pa *PeeringAcceptor) Secret() *Secret {
 
 func (pa *PeeringAcceptor) SecretRef() *SecretRefStatus {
 	return pa.Status.SecretRef
+}
+func (pa *PeeringAcceptor) KubeKind() string {
+	return PeeringAcceptorKubeKind
+}
+func (pa *PeeringAcceptor) KubernetesName() string {
+	return pa.ObjectMeta.Name
+}
+func (pa *PeeringAcceptor) Validate() error {
+	var errs field.ErrorList
+	// The nil checks must return since you can't do further validations.
+	if pa.Spec.Peer == nil {
+		errs = append(errs, field.Invalid(field.NewPath("spec").Child("peer"), pa.Spec.Peer, "peer must be specified"))
+		return apierrors.NewInvalid(
+			schema.GroupKind{Group: ConsulHashicorpGroup, Kind: PeeringAcceptorKubeKind},
+			pa.KubernetesName(), errs)
+	}
+	if pa.Spec.Peer.Secret == nil {
+		errs = append(errs, field.Invalid(field.NewPath("spec").Child("peer").Child("secret"), pa.Spec.Peer.Secret, "secret must be specified"))
+		return apierrors.NewInvalid(
+			schema.GroupKind{Group: ConsulHashicorpGroup, Kind: PeeringAcceptorKubeKind},
+			pa.KubernetesName(), errs)
+	}
+	// Currently, the only supported backend is "kubernetes".
+	if pa.Spec.Peer.Secret.Backend != SecretBackendTypeKubernetes {
+		errs = append(errs, field.Invalid(field.NewPath("spec").Child("peer").Child("secret").Child("backend"), pa.Spec.Peer.Secret.Backend, `backend must be "kubernetes"`))
+	}
+	if len(errs) > 0 {
+		return apierrors.NewInvalid(
+			schema.GroupKind{Group: ConsulHashicorpGroup, Kind: PeeringAcceptorKubeKind},
+			pa.KubernetesName(), errs)
+	}
+	return nil
 }