+++ control-plane/api/v1alpha1/peeringdialer_types.go
@@ -1,10 +1,15 @@
 package v1alpha1
 
 import (
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apimachinery/pkg/util/validation/field"
 )
 
 // NOTE: json tags are required.  Any new fields you add must have json tags for the fields to be serialized.
+
+const PeeringDialerKubeKind = "peeringdialers"
 
 func init() {
 	SchemeBuilder.Register(&PeeringDialer{}, &PeeringDialerList{})
@@ -61,4 +66,36 @@ func (pd *PeeringDialer) Secret() *Secret {
 
 func (pd *PeeringDialer) SecretRef() *SecretRefStatus {
 	return pd.Status.SecretRef
+}
+func (pd *PeeringDialer) KubeKind() string {
+	return PeeringDialerKubeKind
+}
+func (pd *PeeringDialer) KubernetesName() string {
+	return pd.ObjectMeta.Name
+}
+func (pd *PeeringDialer) Validate() error {
+	var errs field.ErrorList
+	// The nil checks must return since you can't do further validations.
+	if pd.Spec.Peer == nil {
+		errs = append(errs, field.Invalid(field.NewPath("spec").Child("peer"), pd.Spec.Peer, "peer must be specified"))
+		return apierrors.NewInvalid(
+			schema.GroupKind{Group: ConsulHashicorpGroup, Kind: PeeringDialerKubeKind},
+			pd.KubernetesName(), errs)
+	}
+	if pd.Spec.Peer.Secret == nil {
+		errs = append(errs, field.Invalid(field.NewPath("spec").Child("peer").Child("secret"), pd.Spec.Peer.Secret, "secret must be specified"))
+		return apierrors.NewInvalid(
+			schema.GroupKind{Group: ConsulHashicorpGroup, Kind: PeeringDialerKubeKind},
+			pd.KubernetesName(), errs)
+	}
+	// Currently, the only supported backend is "kubernetes".
+	if pd.Spec.Peer.Secret.Backend != "kubernetes" {
+		errs = append(errs, field.Invalid(field.NewPath("spec").Child("peer").Child("secret").Child("backend"), pd.Spec.Peer.Secret.Backend, `backend must be "kubernetes"`))
+	}
+	if len(errs) > 0 {
+		return apierrors.NewInvalid(
+			schema.GroupKind{Group: ConsulHashicorpGroup, Kind: PeeringDialerKubeKind},
+			pd.KubernetesName(), errs)
+	}
+	return nil
 }